+++ fiddles/nativeScript/fiddle-0011-SqliteNg2/puzzle/app/app.routing.js
@@ -1,9 +1,12 @@
 "use strict";
 var game_component_1 = require("./pages/game/game.component");
+var about_component_1 = require("./pages/about/about.component");
 exports.routes = [
-    { path: "", component: game_component_1.GameComponent }
+    { path: "", component: game_component_1.GameComponent },
+    { path: "about", component: about_component_1.AboutComponent }
 ];
 exports.navigatableComponents = [
-    game_component_1.GameComponent
+    game_component_1.GameComponent,
+    about_component_1.AboutComponent
 ];
 //# sourceMappingURL=app.routing.js.map