+++ fiddles/nativeScript/fiddle-0011-SqliteNg2/puzzle/app/app.routing.ts
@@ -1,9 +1,12 @@
 import { GameComponent } from "./pages/game/game.component";
+import { AboutComponent } from "./pages/about/about.component";
 
 export const routes = [
-    { path: "", component: GameComponent }
+    { path: "", component: GameComponent },
+    { path: "about", component: AboutComponent }
 ];
 
 export const navigatableComponents = [
-    GameComponent
+    GameComponent,
+    AboutComponent
 ];