+++ fiddles/nativeScript/fiddle-0011-SqliteNg2/puzzle/app/pages/game/game.component.ts
@@ -1,8 +1,11 @@
-import { Component, ElementRef, OnInit, ViewChild } from "@angular/core";
-import { View } from "ui/core/view";
-import { Router } from "@angular/router";
-import { Page } from "ui/page";
-import { Color } from "color";
+import {Component, ElementRef, OnInit, ViewChild} from "@angular/core";
+import {View} from "ui/core/view";
+import {Router} from "@angular/router";
+import {Page} from "ui/page";
+import {Color} from "color";
+
+import { Config } from '../../shared/config';
+
 
 @Component({
     selector: "my-app",
@@ -13,13 +16,13 @@ export class GameComponent implements OnInit {
     @ViewChild("container") container: ElementRef;
     highscoreVisibility: string;
     title: string;
-    
-    constructor(private router: Router, private page: Page) {}
+
+    constructor(private router: Router, private page: Page) {
+    }
 
     ngOnInit() {
-        this.page.actionBarHidden = true;
         this.highscoreVisibility = "collapsed";
-        this.title = "15 Puzzle";
+        this.title = Config.title;
     }
 
     onPlayTap() {
@@ -27,7 +30,7 @@ export class GameComponent implements OnInit {
     }
 
     onAboutTap() {
-
+        this.router.navigate(["/about"]);
     }
 
     onHighScoreTap() {