+++ fiddles/nativeScript/fiddle-0011-SqliteNg2/puzzle/app/shared/config.js
@@ -1,0 +1,10 @@
+"use strict";
+var Config = (function () {
+    function Config() {
+    }
+    return Config;
+}());
+Config.title = "15 Puzzle";
+Config.isDev = false;
+exports.Config = Config;
+//# sourceMappingURL=config.js.map