+++ fiddles/nativeScript/fiddle-0011-SqliteNg2/puzzle/app/shared/config.ts
@@ -1,0 +1,4 @@
+export class Config {
+    static title: string = "15 Puzzle";
+    static isDev: boolean = false;
+}