+++ scripts/bin/javac/_start.sh
@@ -23,7 +23,10 @@ function javacBuild() {
         rm -rf ${__TARGET_DIR__};
     fi
     mkdir ${__TARGET_DIR__};
-    ${__JAVA_COMPILER__} -d ${__TARGET_DIR__} ${__SOURCE_FILE__};
+
+    # javac -sourcepath src -classpath classes;lib\Banners.jar \src\farewells\GoodBye.java -d classes
+
+    ${__JAVA_COMPILER__} -sourcepath ${__SOURCE_PATH__} -classpath ${__CLASS_PATH__} ${__SOURCE_FILE__} -d ${__TARGET_DIR__};
 }
 
 function javaStart() {